refactor AppButton from jsx to tsx
